[FEAT] Implement chatbot command structure; add setup, settings, delete, and info handlers
